* [MM-21921] Adds the server version check before running a command
Before running any command that uses the Client instance, mmctl will
now check that the server version matches its own, and will show an
error if it doesn't.

This PR enables as well the --strict flag and the MMCTL_STRICT
environment variable; if one of those is set, mmctl will error in case
the server version doesn't match and will not run the command.

* Update README.md

Fixed typo

* Update docs

Co-authored-by: lindy65 <lindy65@users.noreply.github.com>