fix ReadFile unused function error
It seems like golangci-cli throws an error when I define a function and use it by assigning it to a variable (to override another function). This is perceived by the linter as not using the function and throws an error.
Therefore, the workaround is to directly assign an anonymous function to the variable.