MM-24294: Check the type before assigning a value to a slice
(reflect.Value).Set was being called without checking whether the input
value's type was indeed a slice or not. This would cause a panic in case
when the user is unsure of how to enter an array value to set a config key.

To prevent this, we check for the type and raise an appropriate error
and also add an example to show how to set an array value.