MM-38701: Automatically retry on a failed export download (#524)
We had a resume functionality which would download the file
again from a given offset. But clients would create wrapper
scripts which would auto-retry.

Therefore, we add that functionality in-built and deprecate
the resume flag. The reasoning behind removing the flag is that
if it doesn't succeed within 5 tries, then probably it's better
to attempt to download the file from scratch again sometime later.

https://mattermost.atlassian.net/browse/MM-38701